requirements.txt
django==1.10
djangorestframework==3.5.3
django-rest-swagger==2.1.1
django-filter==1.0.4
